Move maybe_gzip_open and path_to_str to utils
