+++ json_lines/_lib.py
@@ -1,21 +1,22 @@
 from __future__ import absolute_import
 from contextlib import contextmanager
+
+
 try:
     import ujson as json
 except ImportError:
     import json
 import json
-import gzip
 import logging
 import six
 
+from .utils import maybe_gzip_open
 from ._gzip import recover, get_known_read_position
 
 
 __all__ = ['open', 'reader']
 
 
-builtin_open = open
 logger = logging.getLogger("json_lines")
 
 
@@ -45,21 +46,6 @@ def reader(file, broken=False):
         return _iter_json_lines(file)
     else:
         return _iter_json_lines_recovering(file)
-
-
-def maybe_gzip_open(path, *args, **kwargs):
-    """
-    Open file with either open or gzip.open, depending on file extension.
-
-    This function doesn't handle json lines format, just opens a file
-    in a way it is decoded transparently if needed.
-    """
-    path = path_to_str(path)
-    if path.endswith('.gz') or path.endswith('.gzip'):
-        _open = gzip.open
-    else:
-        _open = builtin_open
-    return _open(path, *args, **kwargs)
 
 
 def _iter_json_lines(file):
@@ -130,15 +116,3 @@ class _JlRecoveringReader(object):
     def _mark_good(self, file, buffered):
         self.last_good_position = get_known_read_position(file.fileobj,
                                                           buffered)
-
-
-def path_to_str(path):
-    """ Convert pathlib.Path objects to str; return other objects as-is. """
-    try:
-        from pathlib import Path as _Path
-    except ImportError:  # Python < 3.4
-        class _Path:
-            pass
-    if isinstance(path, _Path):
-        return str(path)
-    return path
